requirements/dev.txt
# ─────────────────────────────────────────────
#  Requirements DÉVELOPPEMENT uniquement
#  Étend base.txt — NE PAS installer en production
# ─────────────────────────────────────────────
-r base.txt

# Debug toolbar (expose SQL, config Django, variables d'env)
django-debug-toolbar==4.4.6

# Shell amélioré, commandes utilitaires (shell_plus, export_emails, etc.)
# DANGER en prod : accès complet aux modèles via shell
django-extensions==3.2.3
